move store copy command into commands/store/copy/index.ts to prepare for new subcommand alongside it!
