+++ packages/store/src/commands/store/copy/index.test.ts
@@ -1,9 +1,9 @@
-import Copy from './copy.js'
-import {StoreCopyOperation} from '../../services/store/operations/store-copy.js'
-import {StoreExportOperation} from '../../services/store/operations/store-export.js'
-import {StoreImportOperation} from '../../services/store/operations/store-import.js'
-import {ApiClient} from '../../services/store/api/api-client.js'
-import {MockApiClient} from '../../services/store/mock/mock-api-client.js'
+import Copy from './index.js'
+import {StoreCopyOperation} from '../../../services/store/operations/store-copy.js'
+import {StoreExportOperation} from '../../../services/store/operations/store-export.js'
+import {StoreImportOperation} from '../../../services/store/operations/store-import.js'
+import {ApiClient} from '../../../services/store/api/api-client.js'
+import {MockApiClient} from '../../../services/store/mock/mock-api-client.js'
 import {describe, vi, expect, test, beforeEach} from 'vitest'
 import {Config, loadHelpClass} from '@oclif/core'
