+++ packages/store/src/commands/store/copy/index.ts
@@ -1,13 +1,13 @@
-import {BaseBDCommand} from '../../lib/base-command.js'
-import {commonFlags, storeFlags, fileFlags, resourceConfigFlags} from '../../lib/flags.js'
-import {FlagOptions} from '../../lib/types.js'
-import {parseResourceConfigFlags} from '../../lib/resource-config.js'
-import {OperationMode} from '../../services/store/types/operations.js'
-import {StoreCopyOperation} from '../../services/store/operations/store-copy.js'
-import {StoreExportOperation} from '../../services/store/operations/store-export.js'
-import {StoreImportOperation} from '../../services/store/operations/store-import.js'
-import {ApiClient} from '../../services/store/api/api-client.js'
-import {MockApiClient} from '../../services/store/mock/mock-api-client.js'
+import {BaseBDCommand} from '../../../lib/base-command.js'
+import {commonFlags, storeFlags, fileFlags, resourceConfigFlags} from '../../../lib/flags.js'
+import {FlagOptions} from '../../../lib/types.js'
+import {parseResourceConfigFlags} from '../../../lib/resource-config.js'
+import {OperationMode} from '../../../services/store/types/operations.js'
+import {StoreCopyOperation} from '../../../services/store/operations/store-copy.js'
+import {StoreExportOperation} from '../../../services/store/operations/store-export.js'
+import {StoreImportOperation} from '../../../services/store/operations/store-import.js'
+import {ApiClient} from '../../../services/store/api/api-client.js'
+import {MockApiClient} from '../../../services/store/mock/mock-api-client.js'
 import {globalFlags} from '@shopify/cli-kit/node/cli'
 import {joinPath, cwd} from '@shopify/cli-kit/node/path'
 import {loadHelpClass} from '@oclif/core'
